Split eloquent methods
